Revert centralized test gates
