+++ test/gpu/opencl/test_kernel_abstractions.jl
@@ -3,12 +3,22 @@
     include("../common/test_platform.jl")
 
     import pocl_jll
-    using OpenCL: CLArray, cl.finish, cl.queue
+    using OpenCL: CLArray, cl.devices, cl.platforms, cl.finish, cl.queue
     const AT = CLArray
 
-    # TODO: Revisit this once the POCL code generation issues are resolved.
-    block_sizes = fill(256, round_count)
-    synchronize() = finish(queue())
-    test_platform(AT, synchronize)
+    const can_run = any(
+        length(devices(platform)) > 0 for platform in platforms()
+            )
+
+    @testset "Device availability" begin
+        @test can_run
+    end
+
+    if can_run
+        # TODO: Revisit this once the POCL code generation issues are resolved.
+        block_sizes = fill(256, round_count)
+        synchronize() = finish(queue())
+        test_platform(AT, synchronize)
+    end
 
 end
