+++ test/gpu/rocm/test_kernel_abstractions.jl
@@ -2,9 +2,17 @@
 
     include("../common/test_platform.jl")
 
-    using AMDGPU: ROCArray, synchronize
+    using AMDGPU: ROCArray, devices, synchronize
     const AT = ROCArray
 
-    test_platform(AT, synchronize)
+    const can_run = length(devices()) > 0
+
+    @testset "Device availability" begin
+        @test can_run
+    end
+
+    if can_run
+        test_platform(AT, synchronize)
+    end
 
 end
