+++ test/oscar/test_ecc_generalized_toric_codes.jl
@@ -1,4 +1,5 @@
-@testitem "ECC Generalized Toric Code" tags=[:ecc, :ecc_bespoke_checks, :oscar_required] begin
+@testitem "ECC Generalized Toric Code" tags=[:ecc, :ecc_bespoke_checks] begin
+    @static if !Sys.iswindows() && Sys.ARCH == :x86_64 && VERSION >= v"1.11"
         using Oscar
         using QECCore
         import HiGHS
@@ -313,4 +314,5 @@
             c = GeneralizedToric(f, g, α1, α2)
             @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 10
         end
+    end
 end
