+++ test/test_closure_boxes.jl
@@ -1,6 +1,8 @@
-@testitem "Closure boxes" tags=[:julia_1_14_required] begin
-    using Hecke
-    hecke_ext = Base.get_extension(QuantumClifford, :QuantumCliffordHeckeExt)
-    @test !isnothing(hecke_ext)
-    @test isempty(Test.detect_closure_boxes(QuantumClifford, hecke_ext))
+@testitem "Closure boxes" begin
+    @static if VERSION >= v"1.14"
+        using Hecke
+        hecke_ext = Base.get_extension(QuantumClifford, :QuantumCliffordHeckeExt)
+        @test !isnothing(hecke_ext)
+        @test isempty(Test.detect_closure_boxes(QuantumClifford, hecke_ext))
+    end
 end
